refactor: add core/files with traversal-safe resolveSavePath (+tests)
